% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Bartlett}
\alias{Bartlett}
\title{Bartlett kernel function}
\usage{
Bartlett(x)
}
\arguments{
\item{x}{A single numeric.}
}
\value{
A single numeric between 0 and 1.
}
\description{
Evaluate the Bartlett kernel function: \eqn{Bartlett(x)=1-|x|} if
\eqn{|x|\leq 1} and \eqn{Bartlett(x)=1-|x|} otherwise.
}
\examples{

Bartlett(0.5)

}
