% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.multi_result}
\alias{summary.multi_result}
\title{Print the relative importance ratios}
\usage{
\method{summary}{multi_result}(object, ...)
}
\arguments{
\item{object}{An S3 object of class 'multi_result' created by multilevel().}

\item{...}{Additional arguments.}
}
\value{
A matrix containing the summary of the model.
}
\description{
Print the relative importance ratios
}
\examples{

panel <- UKhouse # load the data
est_multi <- multilevel(panel, ic = "BIC3", standarise = TRUE, r_max = 5,
                           depvar_header = "dlPrice", i_header = "Region",
                           j_header = "LPA_Type", t_header = "Date")
summary(est_multi)
}
